Добавил доменные события для корня агрегата Category, написал тесты для проверки их публикации. Добавил поле Currency в ValueObject Money.

+++ backend/ProductService/ProductService.Domain/Entities/Product.cs
@@ -90,7 +90,7 @@ public class Product : IEquatable<Product>
     public static Product Reconstruct(
         Guid id,
         Guid sellerId,
-        int sku,
+        long sku,
         string name,
         string description,
         Money price,
